Split serialization package into api and impl
