+++ src/main/java/me/zeroeightsix/fiber/impl/serialization/JanksonSerializer.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.serialization;
+package me.zeroeightsix.fiber.impl.serialization;
 
 import blue.endless.jankson.Jankson;
 import blue.endless.jankson.JsonElement;
@@ -6,6 +6,8 @@ import blue.endless.jankson.JsonObject;
 import blue.endless.jankson.api.SyntaxError;
 import me.zeroeightsix.fiber.Identifier;
 import me.zeroeightsix.fiber.api.exception.FiberException;
+import me.zeroeightsix.fiber.api.serialization.Marshaller;
+import me.zeroeightsix.fiber.api.serialization.Serializer;
 import me.zeroeightsix.fiber.tree.*;
 
 import javax.annotation.Nonnull;
