+++ src/test/java/me/zeroeightsix/fiber/impl/serialization/JanksonSerializerTest.java
@@ -1,9 +1,9 @@
-package me.zeroeightsix.fiber;
+package me.zeroeightsix.fiber.impl.serialization;
 
+import me.zeroeightsix.fiber.NodeOperationsTest;
 import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
 import me.zeroeightsix.fiber.api.constraint.CompositeType;
 import me.zeroeightsix.fiber.api.exception.FiberException;
-import me.zeroeightsix.fiber.serialization.JanksonSerializer;
 import me.zeroeightsix.fiber.tree.ConfigGroupImpl;
 import me.zeroeightsix.fiber.tree.PropertyMirror;
 import org.junit.jupiter.api.DisplayName;
